HW13 fix after review

+++ hw12_13_14_15_calendar/internal/storage/pgsql/storage.go
@@ -3,7 +3,6 @@ package sqlstorage
 import (
 	"context"
 	"errors"
-	"fmt"
 	"sync"
 	"time"
 
@@ -75,6 +74,14 @@ func (s *storage) CreateEvent(ctx context.Context, event internalStorage.Event) 
 }
 
 func (s *storage) UpdateEvent(ctx context.Context, event internalStorage.Event) error {
+	isExist, err := s.isExistByID(ctx, event.ID)
+	if err != nil {
+		return err
+	}
+	if !isExist {
+		return internalStorage.ErrEventNotFound
+	}
+
 	isBusy, err := s.isDateBusy(ctx, event)
 	if err != nil {
 		return err
@@ -104,9 +111,17 @@ func (s *storage) UpdateEvent(ctx context.Context, event internalStorage.Event) 
 }
 
 func (s *storage) DeleteEvent(ctx context.Context, id string) error {
+	isExist, err := s.isExistByID(ctx, id)
+	if err != nil {
+		return err
+	}
+	if !isExist {
+		return internalStorage.ErrEventNotFound
+	}
+
 	sql := `DELETE FROM events WHERE id=$1`
 
-	_, err := s.conn.Exec(ctx, sql, id)
+	_, err = s.conn.Exec(ctx, sql, id)
 
 	return err
 }
@@ -173,7 +188,6 @@ func (s *storage) EventsByDates(
 
 func (s *storage) isDateBusy(ctx context.Context, event internalStorage.Event) (bool, error) {
 	sql := `SELECT id from events WHERE (start_at BETWEEN $1 AND $2 OR end_at BETWEEN $1 AND $2) AND id != $3`
-	fmt.Println(event.EndAt, event.StartAt, event.ID)
 	row := s.conn.QueryRow(ctx, sql, event.StartAt, event.EndAt, event.ID)
 
 	var id string
@@ -183,4 +197,13 @@ func (s *storage) isDateBusy(ctx context.Context, event internalStorage.Event) (
 	}
 
 	return true, err
+}
+
+func (s *storage) isExistByID(ctx context.Context, id string) (bool, error) {
+	sql := `SELECT EXISTS(SELECT 1 FROM events WHERE id = $1)`
+	row := s.conn.QueryRow(ctx, sql, id)
+	isExist := false
+
+	err := row.Scan(&isExist)
+	return isExist, err
 }
